gjk/requirements.txt
numpy==1.15.2
pygame==1.9.4
PyOpenGL==3.1.1a1